:recycle: Extract ProjectHierarchyDao from ProjectDao to separate responsibilities
