+++ modules/app-server/dao/src/test/kotlin/dao/project/ProjectDaoTest.kt
@@ -206,114 +206,4 @@ internal class ProjectDaoTest : AbstractDaoTest() {
             expectThat(projectDao.findOneById(insertedId2)).isNotNull()
         }
     }
-
-    @Nested
-    inner class TestFindProjectHierarchies {
-        @Test
-        fun `should load project linked modules and environments`() {
-            // Given
-            withProjectInDb(
-                project = ProjectCreationRequest(name = "myProject", "mp"),
-                modulesNames = listOf("mySecondModule"),
-                environments = listOf(EnvironmentCreationRequestWithoutForeignKey(name = "myFirstEnvironment", smallName = "mfe"))
-            )
-
-            // When
-            val projectHierarchies = projectDao.findProjectHierarchies()
-
-            // Then
-            expectThat(
-                projectHierarchies[0]
-            ) {
-                get { name }.isEqualTo("myProject")
-                get { smallName }.isEqualTo("mp")
-                get { envs[0].name }.isEqualTo("myFirstEnvironment")
-                get { envs[0].smallName }.isEqualTo("mfe")
-                get { envs[0].modules[0].name }.isEqualTo("mySecondModule")
-            }
-        }
-
-        @Test
-        fun `should put all modules into each environment hierarchy`() {
-            // Given
-            withProjectInDb(
-                projectName = "myProject",
-                modulesNames = listOf("mySecondModule", "myFirstModule"),
-                environmentsNames = listOf("myFirstEnvironment", "mySecondEnvironment")
-            )
-
-            // When
-            val projectHierarchies = projectDao.findProjectHierarchies()
-
-            // Then
-            expectThat(
-                projectHierarchies[0].envs
-            ).all {
-                get { modules.map { it.name }}.containsExactly("myFirstModule", "mySecondModule")
-            }
-        }
-
-        @Test
-        fun `should sort modules by their names with alphabetical ascendant order`() {
-            // Given
-            withProjectInDb(
-                projectName = "myProject",
-                modulesNames = listOf("mySecondModule", "myFirstModule"),
-                environmentsNames = listOf("myFirstEnvironment")
-            )
-
-            // When
-            val projectHierarchies = projectDao.findProjectHierarchies()
-
-            // Then
-            expectThat(
-                projectHierarchies[0].envs[0].modules.map { it.name }
-            ).containsExactly("myFirstModule", "mySecondModule")
-        }
-
-        @Test
-        fun `should sort environments by their names with alphabetical ascendant order`() {
-            // Given
-            withProjectInDb(
-                projectName = "myProject",
-                environmentsNames = listOf("mySecondEnvironment", "myFirstEnvironment")
-            )
-
-            // When
-            val projectHierarchies = projectDao.findProjectHierarchies()
-
-            // Then
-
-            expectThat(
-                projectHierarchies[0]
-            ) {
-                get { envs[0].name }.isEqualTo("myFirstEnvironment")
-                get { envs[1].name }.isEqualTo("mySecondEnvironment")
-            }
-        }
-
-        @Test
-        fun `should sort projects by their names with alphabetical ascendant order`() {
-            // Given
-            withProjectInDb(
-                projectName = "myProject"
-            )
-
-            withProjectInDb(
-                projectName = "myOtherProject"
-            )
-
-            // When
-            val projectHierarchies = projectDao.findProjectHierarchies()
-
-            // Then
-
-            expectThat(
-                projectHierarchies
-            ) {
-                get(0).get { name }.isEqualTo("myOtherProject")
-                get(1).get { name }.isEqualTo("myProject")
-            }
-        }
-    }
 }
