+++ modules/app-server/domain/src/main/kotlin/project/ProjectHierarchyDaoInterface.kt
@@ -1,0 +1,5 @@
+package project
+
+interface ProjectHierarchyDaoInterface {
+    fun findProjectHierarchies(): List<ProjectHierarchy>
+}
