+++ modules/app-server/domain/src/main/kotlin/project/ProjectService.kt
@@ -2,12 +2,15 @@ package project
 
 import java.util.*
 
-class ProjectService(private val projectDao: ProjectDaoInterface) {
+class ProjectService(
+    private val projectDao: ProjectDaoInterface,
+    private val projectHierarchyDao: ProjectHierarchyDaoInterface
+) {
     fun insert(data: ProjectCreationRequest): Project = projectDao.insert(data)
     fun updateProjectName(id: UUID, data: ProjectNameUpdateRequest): Project? =
         projectDao.updateProjectName(id, data)
 
     fun delete(id: UUID) = projectDao.delete(id)
     fun findOneById(id: UUID): Project? = projectDao.findOneById(id)
-    fun findProjectHierarchies(): List<ProjectHierarchy> = projectDao.findProjectHierarchies()
+    fun findProjectHierarchies(): List<ProjectHierarchy> = projectHierarchyDao.findProjectHierarchies()
 }
